feat(ios): share Active/Upcoming alerts tabs between station and route views
Extract the segmented Active/Upcoming service alerts section from
RouteStatusView into a standalone ServiceAlertsSection component (also
moves ServiceAlertFilter and ServiceAlertCard) and adopt it in
StationDetailsView so the station details screen now defaults to
currently-active alerts with an Upcoming tab, matching route details.
